# lumen/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mario Baggio test",
        "contact": {
            "name": "Mario Baggio",
            "email": "mario.baggio@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/merchants/{name}/{page_number}": {
            "get": {
                "tags": [
                    "Merchants"
                ],
                "description": "Merchants list",
                "operationId": "App\\Http\\Controllers\\MerchantsController::getMerchants",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Merchant name",
                        "required": false,
                        "allowEmptyValue": true,
                        "example": "Ebay"
                    },
                    {
                        "name": "page_number",
                        "in": "path",
                        "description": "Page number",
                        "required": false,
                        "allowEmptyValue": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Merchants list"
                    }
                }
            }
        },
...
        "/sitemap_categories/{sitemap_category_id}/merchants/{page_number}": {
            "get": {
                "tags": [
                    "Sitemap categories"
                ],
                "description": "Sitemap cateogory - merchants list",
                "operationId": "App\\Http\\Controllers\\SitemapCategoriesController::getMerchants",
                "parameters": [
                    {
                        "name": "sitemap_category_id",
                        "in": "path",
                        "description": "Sitemap category id",
                        "required": true,
                        "example": 12
                    },
                    {
                        "name": "page_number",
                        "in": "path",
                        "description": "Page number",
                        "required": false,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sitemap cateogory - merchants list"
                    }
                }
            }
        },
        "/sitemap_categories/{name}/{parent_id}": {
            "post": {
                "tags": [
                    "Sitemap categories"
                ],
                "description": "New Sitemap Category",
                "operationId": "App\\Http\\Controllers\\SitemapCategoriesController::create",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Category name",
                        "required": true,
                        "example": "NY News papers"
                    },
                    {
                        "name": "parent_id",
                        "in": "path",
                        "description": "Category parent id",
                        "required": false,
                        "allowEmptyValue": true,
                        "example": 12
                    }
                ],
                "responses": {
                    "201": {
                        "description": "New Sitemap cateogory addedd"
                    },
                    "412": {
                        "description": "Precondition Failed"
                    }
                }
            }
        },
        "/sitemap_categories/{id}/{new_name}": {
            "patch": {
                "tags": [
                    "Sitemap categories"
                ],
                "description": "Update Sitemap Category name",
                "operationId": "App\\Http\\Controllers\\SitemapCategoriesController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "example": "1"
                    },
                    {
                        "name": "new_name",
                        "in": "path",
                        "description": "Category new name",
                        "required": true,
                        "example": "NY News Paper"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sitemap cateogory updated"
                    },
                    "412": {
                        "description": "Precondition Failed"
                    }
                }
            }
        },
        "/sitemap_categories/{id}": {
            "delete": {
                "tags": [
                    "Sitemap categories"
                ],
                "description": "Delete Sitemap Category",
                "operationId": "App\\Http\\Controllers\\SitemapCategoriesController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sitemap cateogory deleted"
                    },
                    "412": {
                        "description": "Precondition Failed"
                    }
                }
            }
        }
    }
}